[chore] #169 - ProfileType enum 타입 파일 분리
